Move ScanNetworkTask to upper level
so it can be used by others
